+++ Documentation/changelog.rst
@@ -11,6 +11,13 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.1.0...HEAD>`_
 --------------------------------------------------------------------------------------------
 
+Added
+^^^^^
+
+
+* Common translations for depending extensions
+* Support for Croatian in JobRouterLanguageResolver
+
 `1.1.0 <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.0.0...v1.1.0>`_ - 2021-11-21
 ------------------------------------------------------------------------------------------------------