+++ Documentation/changelog.rst
@@ -11,6 +11,12 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.2.0...HEAD>`_
 --------------------------------------------------------------------------------------------
 
+Removed
+^^^^^^^
+
+
+* Compatibility with PHP 7.3
+
 `1.2.0 <https://github.com/brotkrueml/typo3-jobrouter-base/compare/v1.1.0...v1.2.0>`_ - 2022-01-18
 ------------------------------------------------------------------------------------------------------